Get integration tests working.
