+++ tests/integration/components/submission-action-cell-test.js
@@ -11,7 +11,7 @@ test('it renders', function (assert) {
 
   this.render(hbs`{{submission-action-cell}}`);
 
-  assert.equal(this.$().text().trim(), '');
+  assert.equal(this.$().text().trim(), 'No actions available.');
 
   // Template block usage:
   this.render(hbs`
@@ -20,5 +20,5 @@ test('it renders', function (assert) {
     {{/submission-action-cell}}
   `);
 
-  assert.equal(this.$().text().trim(), 'template block text');
+  assert.equal(this.$().text().trim(), 'No actions available.');
 });
