+++ tests/integration/components/workflow-metadata-test.js
@@ -1,15 +1,29 @@
-import { moduleForComponent, test } from 'ember-qunit';
+import { moduleForComponent, test, setupRenderingTest } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
+import { module } from 'qunit';
+import { render } from '@ember/test-helpers';
 
-moduleForComponent('workflow-metadata', 'Integration | Component | workflow metadata', {
-  integration: true
-});
 
-test('it renders', function (assert) {
-  // Set any properties with this.set('myProperty', 'value');
-  // Handle any actions with this.on('myAction', function(val) { ... });
+module('Integration | Component | workflow metadata', (hooks) => {
+  setupRenderingTest(hooks);
+  test('it renders', function (assert) {
+    let model = {};
 
-  // Template usage:
-  this.render(hbs`{{workflow-metadata}}`);
-  assert.ok(true);
+    // TODO: add actual tests here
+    model.newSubmission = Ember.Object.create({
+      repositories: [],
+      grants: []
+    });
+    model.metadataForms = [Ember.Object.create()];
+
+    let doiInfo = Ember.Object.create();
+
+    this.set('model', model);
+    this.set('doiInfo', doiInfo);
+
+    console.log('hi there');
+
+    render(hbs`{{workflow-metadata model=model doiInfo=doiInfo}}`);
+    assert.ok(true);
+  });
 });
